dw-hdmi-audio: ensure that we sync with the irq before removing buffer
Fix a race between the interrupt handler and tearing down the buffer
mapping.  It is possible for the interrupt handler to still be running
on another CPU after we clear the mask register, and we can get as far
as unmapping the buffer.  This causes the CPU running the interrupt
handler to oops.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>